requirements.txt
pytelegrambotapi
youtube-search-python
yt-dlp
Pillow
requests